This is the Amazon Project
I've analysis in Amazon prime data. Where i drop 2 columns at first and then check for duplicate and describe the data and in the meantime, I feel all the N/A values. After that i check 10 directors with most directed movies, whereas we count the each of content type (Movie & TV shows) and the rest of the thing is in the file.